% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{add_analysis}
\alias{add_analysis}
\title{add_analysis}
\usage{
add_analysis(meta, name, description)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{name}{A name for the analysis.}

\item{description}{A sentence describing the aim of the analysis.}
}
\value{
Returns an object of class \code{fgcMeta}. Also saves data in \code{analysis.csv} in \code{meta$data_dir}.
}
\description{
Adds an analysis to an object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}. If an analysis already exists, nothing is done as only a single analysis can be associated with a given \code{fgcMeta} object.
}
