% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_general.R
\name{add_general}
\alias{add_general}
\title{add_general}
\usage{
add_general(
  meta,
  user_email,
  process_methoda = "basep",
  process_methodn = "base"
)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{user_email}{An email address.}

\item{process_methoda}{A character string, defaults to "basep".}

\item{process_methodn}{A character string, defaults to "base".}
}
\description{
Adds general information for use in building an analysis config JSON file.
}
