% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta.R
\name{add_meta}
\alias{add_meta}
\title{add_meta}
\usage{
add_meta(meta)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}
}
\value{
Returns an object of class \code{fgcMeta}. Also saves data in \code{sample.csv} in \code{meta$data_dir}.
}
\description{
Adds a \code{meta} section (after all comparisons have been added) to an object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.
}
