% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_comparison_sample_class.R
\name{edit_comparison_sample_class}
\alias{edit_comparison_sample_class}
\title{edit_comparison_sample_class}
\usage{
edit_comparison_sample_class(meta, name, samples, new_class)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{name}{A name for the comparison.}

\item{samples}{\code{sequenced_sample} IDs for the samples whose class is to be changed.}

\item{new_class}{The new class.}
}
\value{
Returns an object of class \code{fgcMeta}. Also saves data to \code{comparison_sample.csv} in \code{meta$data_dir}.
}
\description{
Make changes to the \code{class} of specific comparison samples to allow bespoke analysis comparisons.
}
