% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_analysis_config_JSON.R
\name{make_analysis_config_JSON}
\alias{make_analysis_config_JSON}
\title{make_analysis_config_JSON}
\usage{
make_analysis_config_JSON(meta, file)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{file}{A file name to write the config JSON to.}
}
\value{
Writes a JSON file to disk.
}
\description{
Writes a pipeline analysis config JSON file for an object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.
}
