% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fgcMeta.R
\name{print.fgcMeta}
\alias{print.fgcMeta}
\title{print.fgcMeta}
\usage{
\method{print}{fgcMeta}(x, ...)
}
\arguments{
\item{x}{An object of classs \code{fgcMeta}.}

\item{...}{Arguments to be passed to \code{print}.}
}
\value{
Prints a summary to the console.
}
\description{
Prints a summary of an \code{fgcMeta} object to the console.
}
