% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ci_submission.R
\name{read_ci_submission}
\alias{read_ci_submission}
\title{read_ci_submission}
\usage{
read_ci_submission(file)
}
\arguments{
\item{file}{A path to a CIGC sequencing submission \code{.xlsx} file.}
}
\value{
A named list containing the header information in \code{header} and a data frame in \code{samples} containing the samples and their indices.
}
\description{
Read in data from a CIGC sequencing submission spreadsheet.
}
