% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unused_samples.R
\name{remove_unused_samples}
\alias{remove_unused_samples}
\title{remove_unused_samples}
\usage{
remove_unused_samples(meta)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}
}
\value{
Returns an object of class \code{fgcMeta}.
}
\description{
Remove any samples that are not used in any comparisons.
}
