% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing_metadata.R
\name{sequencing_metadata}
\alias{sequencing_metadata}
\title{sequencing_metadata}
\usage{
sequencing_metadata(meta, ci_xlsx)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{ci_xlsx}{A path to a CI sequencing request \code{.xlsx} file.}
}
\description{
Add sample metadata to an object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.
}
