% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_fgc_project.R
\name{start_fgc_project}
\alias{start_fgc_project}
\title{start_fgc_project}
\usage{
start_fgc_project(data_dir, project_id, species, dir_entities = NULL)
}
\arguments{
\item{data_dir}{A path to a directory where the data tables will be saved.}

\item{project_id}{A project name (e.g. FGC_0001).}

\item{species}{A species name.}

\item{dir_entities}{A path to a directory containing entities in a file called \code{entities.rds}.}
}
\value{
Returns an object of class \code{fgcMeta}. Also enters data into a project file.
}
\description{
Enter initial data for a project.
}
