% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_metadata.R
\name{sample_metadata}
\alias{sample_metadata}
\title{sample_metadata}
\usage{
sample_metadata(meta, sample_csv)
}
\arguments{
\item{meta}{An object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.}

\item{sample_csv}{A path to a sample '.csv' file.}
}
\value{
Returns an object of class \code{fgcMeta}. Also saves data in \code{sample.csv} in \code{meta$data_dir}.
}
\description{
Add sample metadata to an object of class \code{fgcMeta}, as produced by \code{fgcMetadata::start_fgc_project}.
}
